VEOSS-1430 | Refactor RegistrationService
